Extract useCoverIsDark hook to a file
